requirements.txt

Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=4.5.0
elasticsearch>=8.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
django-cors-headers>=4.0.0
